refactor: move loadConfig out of rspack-cli
