+++ packages/rspack-cli/src/rspack-cli.ts
@@ -1,27 +1,20 @@
 import { hideBin } from "yargs/helpers";
 import yargs from "yargs";
 import util from "util";
-import path from "path";
-import fs from "fs";
 import { RspackCLIColors, RspackCLILogger, RspackCLIOptions } from "./types";
 import { BuildCommand } from "./commands/build";
 import { ServeCommand } from "./commands/serve";
 import {
 	RspackOptions,
-	MultiCompilerOptions,
-	createMultiCompiler,
-	createCompiler,
 	MultiCompiler,
 	Compiler,
 	rspack,
 	MultiRspackOptions
 } from "@rspack/core";
 import { normalizeEnv } from "./utils/options";
+import { loadRspackConfig } from "./utils/loadConfig";
 import { Mode } from "@rspack/core/src/config";
 
-const defaultConfig = "rspack.config.js";
-const defaultEntry = "src/index.js";
-type Callback<T> = <T>(err: Error, res?: T) => void;
 type RspackEnv = "development" | "production";
 export class RspackCLI {
 	colors: RspackCLIColors;
@@ -181,42 +174,7 @@ export class RspackCLI {
 	async loadConfig(
 		options: RspackCLIOptions
 	): Promise<RspackOptions | MultiRspackOptions> {
-		let loadedConfig:
-			| undefined
-			| RspackOptions
-			| MultiRspackOptions
-			| ((
-					env: Record<string, any>,
-					argv: Record<string, any>
-			  ) => RspackOptions | MultiRspackOptions);
-		// if we pass config paras
-		if (options.config) {
-			const resolvedConfigPath = path.resolve(process.cwd(), options.config);
-			if (!fs.existsSync(resolvedConfigPath)) {
-				throw new Error(`config file "${resolvedConfigPath}" not exists`);
-			}
-			loadedConfig = require(resolvedConfigPath);
-		} else {
-			let defaultConfigPath = path.resolve(process.cwd(), defaultConfig);
-			if (fs.existsSync(defaultConfigPath)) {
-				loadedConfig = require(defaultConfigPath);
-			} else {
-				let entry: Record<string, string> = {};
-				if (options.entry) {
-					entry = {
-						main: options.entry.map(x => path.resolve(process.cwd(), x))[0] // Fix me when entry supports array
-					};
-				} else {
-					entry = {
-						main: path.resolve(process.cwd(), defaultEntry)
-					};
-				}
-				loadedConfig = {
-					entry
-				};
-			}
-		}
-
+		let loadedConfig = loadRspackConfig(options);
 		if (options.configName) {
 			const notFoundConfigNames: string[] = [];
